Tag Propagation Algorithm
This component is designed to perform the chaining of information between all the album images. It uses all the information collected so far, and, while utilizing the graph structure, chains tags between the images, based on the relationships between them. This component receives as input a file containing a list of key images that were suggested to the user for tagging, and a list of tags. The output will be the list of images and the tags that were chained to each, along with a confidence score that represents the system's degree of confidence in the chaining of each tag.